% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerSeqStudy.R
\name{binom.false.pos}
\alias{binom.false.pos}
\title{calculates the false positives in a binomial model
if there is over-diserspion}
\usage{
binom.false.pos(my.alpha, my.beta, N, Leff = 1500 * 3/4, num.genes = 18500,
  signif.level = 5e-06)
}
\arguments{
\item{my.alpha}{alpha parameter for beta binomial}

\item{my.beta}{beta parameter for beta binomial}

\item{N}{vector of # samples to calculate power for}

\item{Leff}{effective gene length in bases}

\item{num.genes}{number of genes that are tested}

\item{signif.level}{alpha level for power analysis}
}
\description{
calculates the false positives in a binomial model
if there is over-diserspion
}

