% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerSeqStudy.R
\name{binom.power}
\alias{binom.power}
\title{calculates the power in a binomial power model}
\usage{
binom.power(my.mu, N, Leff = 1500 * 3/4, r = 0.02, signif.level = 5e-06)
}
\arguments{
\item{my.mu}{per base rate of mutation for binomial}

\item{N}{vector of sample sizes}

\item{r}{effect size for power analysis}

\item{signif.level}{alpha level for power analysis}
}
\value{
vector containing power for each sample size
}
\description{
calculates the power in a binomial power model
}

