% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerSeqStudy.R
\name{rateCvToAlphaBeta}
\alias{rateCvToAlphaBeta}
\title{Converts mutation rate and coefficient of variation (CV) parameters
to equivalent alpha and beta parameters typically used for beta-binomial.}
\usage{
rateCvToAlphaBeta(rate, cv)
}
\arguments{
\item{rate}{mutation rate}

\item{cv}{coefficient of variation for mutation rate}
}
\value{
Param list containing alpha and beta
}
\description{
Converts mutation rate and coefficient of variation (CV) parameters
to equivalent alpha and beta parameters typically used for beta-binomial.
}

