% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioMetric.R
\name{ratiometricBbdPoweredEffectSize}
\alias{ratiometricBbdPoweredEffectSize}
\title{Calculates the effect size of a driver gene according to a beta-binomial model
of ratio-metric features for which there is sufficient power.}
\usage{
ratiometricBbdPoweredEffectSize(possible.effect.sizes, desired.power, p, cv, mu,
  samp.size, Df = 1, signif.level = 5e-06, Leff = 1500 * 3/4)
}
\arguments{
\item{possible.effect.sizes}{vector of effect sizes}

\item{desired.power}{A floating point number indicating desired power}

\item{p}{the background fraction of total mutations represented by the ratio-metric feature (e.g. inactivating mutations / total)}

\item{cv}{the coefficient of variation for the parameter p}

\item{mu}{Mutation rate per base}

\item{samp.size}{number of cancer samples in study}

\item{Df}{fraction of driver mutations that are the specific one of interest}

\item{signif.level}{significance level for binomial test}

\item{Leff}{effective gene length of CDS in bases for an average gene}
}
\value{
List containing the smallest effect size with sufficient power
}
\description{
Effect size is measures as the fraction of sample/patient cancers with a 
mutation in a driver gene above the background mutation rate.
}

