% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioMetric.R
\name{ratiometricBbdRequiredSampleSize}
\alias{ratiometricBbdRequiredSampleSize}
\title{Calculates the smallest sample size to detect driver genes for which
there is sufficient power using a beta-binomial model for ratio-metric features.}
\usage{
ratiometricBbdRequiredSampleSize(p, cv, desired.power, possible.samp.sizes, mu,
  effect.size, Df = 1, signif.lvl = 5e-06, Leff = 1500 * 3/4)
}
\arguments{
\item{p}{the background fraction of total mutations represented by the ratio-metric feature (e.g. inactivating mutations / total)}

\item{cv}{the coefficient of variation for the parameter p}

\item{desired.power}{A floating point number indicating desired power}

\item{possible.samp.sizes}{vector of possible number of cancer samples in study}

\item{mu}{mutation rate per base}

\item{effect.size}{fraction of samples above background mutation rate}

\item{signif.level}{significance level for binomial test}

\item{L}{gene length of CDS in bases for an average gene}
}
\value{
List containing the smallest effect size with sufficient power
}
\description{
Effect size is measures as the fraction of sample/patient cancers with a non-silent
mutation in a driver gene above the background mutation rate.
}

