% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerSeqStudy.R
\name{runRatiometricAnalysis}
\alias{runRatiometricAnalysis}
\title{Runs the entire power and false positive analysis pipeline.}
\usage{
runRatiometricAnalysis(p, mu, effect.size, signif.level, samp.sizes,
  desired.power = 0.9, Leff = 1500 * 3/4, Df = 1, possible.cvs = c())
}
\description{
Runs the entire power and false positive analysis pipeline.
}

