% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerSeqStudy.R
\name{runSmgAnalysis}
\alias{runSmgAnalysis}
\title{Runs the entire power and false positive analysis pipeline.}
\usage{
runSmgAnalysis(mu, effect.size, signif.level, samp.sizes, desired.power = 0.9,
  Leff = 1500 * 3/4, possible.cvs = c())
}
\description{
Runs the entire power and false positive analysis pipeline.
}

