% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smg.R
\name{smg.bbd.power}
\alias{smg.bbd.power}
\title{calculates the power in a beta-binomial model for
significantly mutated genes.}
\usage{
smg.bbd.power(my.alpha, my.beta, N, Leff = 1500 * 3/4, r = 0.02,
  signif.level = 5e-06)
}
\arguments{
\item{my.alpha}{alpha parameter for beta binomial}

\item{my.beta}{beta parameter for beta binomial}

\item{N}{maximum number of sample to calculate power for}

\item{Leff}{effective gene length in bases}

\item{r}{effect size for power analysis}

\item{signif.level}{alpha level for power analysis}
}
\description{
calculates the power in a beta-binomial model for
significantly mutated genes.
}

