% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smg.R
\name{smgBbdPoweredEffectSize}
\alias{smgBbdPoweredEffectSize}
\title{Calculates the smallest effect size in a driver gene for which
there is sufficient power using a significantly mutated gene
approach with a beta-binomial model.}
\usage{
smgBbdPoweredEffectSize(possible.effect.sizes, desired.power, mu, cv, samp.size,
  signif.level = 5e-06, Leff = 1500 * 3/4)
}
\arguments{
\item{possible.effect.sizes}{vector of effect sizes}

\item{desired.power}{A floating point number indicating desired power}

\item{mu}{Mutation rate per base}

\item{cv}{Coefficient of Variation surrounding the uncertaintly in mutation rate}

\item{samp.size}{number of cancer samples in study}

\item{signif.level}{significance level for binomial test}

\item{Leff}{effective gene length of CDS in bases for an average gene}
}
\value{
List containing the smallest effect size with sufficient power
}
\description{
Effect size is measures as the fraction of sample/patient cancers with a non-silent
mutation in a driver gene above the background mutation rate.
}

