% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerSeqStudy.R
\name{thisfile}
\alias{thisfile}
\title{Utility function to get location of script when
executing Rscript.}
\usage{
thisfile()
}
\description{
Utility function to get location of script when
executing Rscript.
}

