% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioMetric.R
\name{ratiometric.bbd.power}
\alias{ratiometric.bbd.power}
\title{Calculates the power in a ratio-metric approach using 
a beta-binomial power model.}
\usage{
ratiometric.bbd.power(my.alpha, my.beta, N, mu, Df = 1, Leff = 1500 * 3/4,
  r = 0.02, signif.level = 5e-06)
}
\arguments{
\item{my.alpha}{alpha parameter for beta binomial}

\item{my.beta}{beta parameter for beta binomial}

\item{N}{vector of sample sizes}

\item{mu}{per base rate of mutation}

\item{Leff}{length of gene in bases}

\item{r}{effect size for power analysis}

\item{signif.level}{alpha level for power analysis}
}
\value{
vector containing power for each sample size
}
\description{
The alpha and beta parameterize a proportion out of the
total mutations in a gene, rather than a mutation rate per base.
}

