% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioMetric.R
\name{ratiometric.binom.false.pos}
\alias{ratiometric.binom.false.pos}
\title{calculates the false positives for a binomial model of
a ratio-metric feature.}
\usage{
ratiometric.binom.false.pos(my.alpha, my.beta, N, mu, Leff = 1500 * 3/4,
  num.genes = 18500, signif.level = 5e-06)
}
\arguments{
\item{my.alpha}{alpha parameter for beta binomial}

\item{my.beta}{beta parameter for beta binomial}

\item{N}{vector of # samples to calculate power for}

\item{mu}{mutation rate per base}

\item{num.genes}{number of genes that are tested}

\item{signif.level}{alpha level for power analysis}

\item{L}{gene CDS length in bases}
}
\description{
calculates the false positives for a binomial model of
a ratio-metric feature.
}

